requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
python-dotenv==0.21.0
yarl==1.8.1
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
colorama==0.4.5
discord.py==2.0.1
frozenlist==1.3.1
halo==0.0.31
idna==3.4
log-symbols==0.0.14
multidict==6.0.2
python-dotenv==0.21.0
six==1.16.0
spinners==0.0.24
termcolor==2.0.1
yarl==1.8.1
